RealStep fix

+++ Syntage.Framework/Tools/Log.cs
@@ -7,22 +7,38 @@ namespace Syntage.Framework.Tools
 {
     public class Log
     {
-        private readonly string _path;
+        public static Log Instance { get; } = new Log();
 
-        private Log()
-        {
-            _path = string.Format("{0}\\{1}{2}",Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), Assembly.GetExecutingAssembly().GetName().Name, "Log.txt");
-
-            if (File.Exists(_path))
-                File.Delete(_path);
-
-            File.WriteAllText(_path, string.Empty);
-        }
+        private string _path;
+        private bool _writeToFile;
 
         public event Action<string> OnLog;
 
-        public static Log Instance { get; } = new Log();
+        private Log()
+        {
+        }
 
+        public bool WriteToFile
+        {
+            get { return _writeToFile; }
+            set
+            {
+                if (value == _writeToFile)
+                    return;
+
+                _writeToFile = value;
+                if (_writeToFile)
+                {
+                    _path = string.Format("{0}\\{1}{2}",Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), Assembly.GetExecutingAssembly().GetName().Name, "Log.txt");
+
+                    if (File.Exists(_path))
+                        File.Delete(_path);
+
+                    File.WriteAllText(_path, string.Empty);                    
+                }
+            }
+        }
+        
         public static void Print(object message)
         {
             Instance.Write(message);
@@ -30,11 +46,11 @@ namespace Syntage.Framework.Tools
 
         private void Write(object message)
         {
-#if LOG_ENABLE
-            File.AppendAllText(_path, string.Format("{0} : {1}{2}", DateTime.Now.ToString(CultureInfo.InvariantCulture), message, Environment.NewLine));
+            if (WriteToFile)
+                File.AppendAllText(_path, string.Format("{0} : {1}{2}",
+                    DateTime.Now.ToString(CultureInfo.InvariantCulture), message, Environment.NewLine));
 
             OnLog?.Invoke(message.ToString());
-#endif
         }
     }
 }
